Updates and Feature Additions
Added spring security, recquiring a login to access survey settings.

When answers are submitted, the survey name is also submitted as a parameter.

Start of results page. Ability to access all entries made.

Modified structure of answer tables. Changed question_id and answer_id column from INT to VARCHAR.

When an answer is submitted it no longer goes to the answers table but instead it's own surveyname_answers table.